Working similarly as the spec mandates
Now it is creating a default application for every deployed standalone
resource. You can also deploy applications. For the moment you can only
register providers to registered applications and the support for the
filters and interceptor with the base url is missing.

Build is still not working properly. Should fix that.
There is an initialization error log but everything works.
Updated to latest CXF version.

Needs cleanup... a lot of code moved around. :-(